chainpulse/backend/requirements.txt
fastapi
uvicorn[standard]
networkx
google-generativeai
pydantic
python-dotenv
websockets
httpx
firebase-admin
